+++ test/OpenTelemetry.Exporter.Zipkin.Tests/Implementation/ZipkinActivityExporterRemoteEndpointTests.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using OpenTelemetry.Exporter.Zipkin.Tests;
-using OpenTelemetry.Trace;
 using Xunit;
 
 namespace OpenTelemetry.Exporter.Zipkin.Implementation.Tests;
@@ -28,10 +27,7 @@ public class ZipkinActivityExporterRemoteEndpointTests
     {
         // Arrange
         using var activity = ZipkinExporterTests.CreateTestActivity(
-            additionalAttributes: new Dictionary<string, object>
-            {
-                ["net.peer.name"] = "RemoteServiceName",
-            });
+            additionalAttributes: new Dictionary<string, object> { ["net.peer.name"] = "RemoteServiceName", });
 
         // Act & Assert
         var zipkinSpan = ZipkinActivityConversionExtensions.ToZipkinSpan(activity, DefaultZipkinEndpoint);
@@ -45,111 +41,13 @@ public class ZipkinActivityExporterRemoteEndpointTests
     public void GenerateActivity_RemoteEndpointResolutionPriority(RemoteEndpointPriorityTestCase testCase)
     {
         // Arrange
-        using var activity = ZipkinExporterTests.CreateTestActivity(additionalAttributes: testCase.RemoteEndpointAttributes!);
+        using var activity =
+            ZipkinExporterTests.CreateTestActivity(additionalAttributes: testCase.RemoteEndpointAttributes!);
 
         // Act & Assert
         var zipkinSpan = ZipkinActivityConversionExtensions.ToZipkinSpan(activity, DefaultZipkinEndpoint);
 
         Assert.NotNull(zipkinSpan.RemoteEndpoint);
         Assert.Equal(testCase.ExpectedResult, zipkinSpan.RemoteEndpoint.ServiceName);
-    }
-
-    public class RemoteEndpointPriorityTestCase
-    {
-        public static TheoryData<RemoteEndpointPriorityTestCase> TestCases =>
-        [
-            new()
-            {
-                Name = "Highest priority name = net.peer.name",
-                ExpectedResult = "RemoteServiceName",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    ["http.host"] = "DiscardedRemoteServiceName",
-                    ["net.peer.name"] = "RemoteServiceName",
-                    ["peer.hostname"] = "DiscardedRemoteServiceName",
-                },
-            },
-            new()
-            {
-                Name = "Highest priority name = SemanticConventions.AttributePeerService",
-                ExpectedResult = "RemoteServiceName",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    [SemanticConventions.AttributePeerService] = "RemoteServiceName",
-                    ["http.host"] = "DiscardedRemoteServiceName",
-                    ["net.peer.name"] = "DiscardedRemoteServiceName",
-                    ["net.peer.port"] = "1234",
-                    ["peer.hostname"] = "DiscardedRemoteServiceName",
-                },
-            },
-            new()
-            {
-                Name = "Only has net.peer.name and net.peer.port",
-                ExpectedResult = "RemoteServiceName:1234",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    ["net.peer.name"] = "RemoteServiceName",
-                    ["net.peer.port"] = "1234",
-                },
-            },
-            new()
-            {
-                Name = "net.peer.port is an int",
-                ExpectedResult = "RemoteServiceName:1234",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    ["net.peer.name"] = "RemoteServiceName",
-                    ["net.peer.port"] = 1234,
-                },
-            },
-            new()
-            {
-                Name = "Has net.peer.name and net.peer.port",
-                ExpectedResult = "RemoteServiceName:1234",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    ["http.host"] = "DiscardedRemoteServiceName",
-                    ["net.peer.name"] = "RemoteServiceName",
-                    ["net.peer.port"] = "1234",
-                    ["peer.hostname"] = "DiscardedRemoteServiceName",
-                },
-            },
-            new()
-            {
-                Name = "Has net.peer.ip and net.peer.port",
-                ExpectedResult = "1.2.3.4:1234",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    ["http.host"] = "DiscardedRemoteServiceName",
-                    ["net.peer.ip"] = "1.2.3.4",
-                    ["net.peer.port"] = "1234",
-                    ["peer.hostname"] = "DiscardedRemoteServiceName",
-                },
-            },
-            new()
-            {
-                Name = "Has net.peer.name, net.peer.ip, and net.peer.port",
-                ExpectedResult = "RemoteServiceName:1234",
-                RemoteEndpointAttributes = new Dictionary<string, object>
-                {
-                    ["http.host"] = "DiscardedRemoteServiceName",
-                    ["net.peer.name"] = "RemoteServiceName",
-                    ["net.peer.ip"] = "1.2.3.4",
-                    ["net.peer.port"] = "1234",
-                    ["peer.hostname"] = "DiscardedRemoteServiceName",
-                },
-            },
-        ];
-
-        public string? Name { get; private set; }
-
-        public string? ExpectedResult { get; private set; }
-
-        public Dictionary<string, object>? RemoteEndpointAttributes { get; private set; }
-
-        public override string? ToString()
-        {
-            return this.Name;
-        }
     }
 }
